# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cargo Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers/profile:
    get:
      operationId: getDriverProfile
      tags: [drivers]
      summary: Get driver profile
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    patch:
      operationId: updateDriverProfile
      tags: [drivers]
      summary: Update driver profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
...
  /drivers/status:
    patch:
      operationId: updateDriverStatus
      tags: [drivers]
      summary: Update driver online/offline status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  # ---- ANALYTICS / DASHBOARD ----
...
  /faq:
    get:
      operationId: listFaq
      tags: [public]
      summary: List published FAQs
      parameters:
        - in: query
          name: categoryId
          schema: { type: integer }
      responses:
        "200":
          description: FAQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqItem"
...
  /admin/faq:
    get:
      operationId: listAdminFaq
      tags: [admin]
      summary: List all FAQs (admin)
      responses:
        "200":
          description: FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqItem"
    post:
      operationId: createAdminFaq
      tags: [admin]
      summary: Create FAQ item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
  /admin/faq/{faqId}:
    patch:
      operationId: updateAdminFaq
      tags: [admin]
      summary: Update FAQ item
      parameters:
        - in: path
          name: faqId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
    delete:
      operationId: deleteAdminFaq
      tags: [admin]
      summary: Delete FAQ item
      parameters:
        - in: path
          name: faqId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
  /admin/faq/categories:
    get:
      operationId: listAdminFaqCategories
      tags: [admin]
      summary: List FAQ categories (admin)
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqCategory"
    post:
      operationId: createAdminFaqCategory
      tags: [admin]
      summary: Create FAQ category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqCategory"
  /admin/faq/categories/{categoryId}:
    patch:
      operationId: updateAdminFaqCategory
      tags: [admin]
      summary: Update FAQ category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqCategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqCategory"
    delete:
      operationId: deleteAdminFaqCategory
      tags: [admin]
      summary: Delete FAQ category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqCategory"

  # ─── ADMIN BLOG ───────────────────────────────────────────────
...
  /admin/flash-sales:
    get:
      operationId: listAdminFlashSales
      tags: [admin]
      summary: List all flash sales (admin)
      responses:
        "200":
          description: Flash sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminFlashSale"
    post:
      operationId: createAdminFlashSale
      tags: [admin]
      summary: Create flash sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashSaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFlashSale"
  /admin/flash-sales/{saleId}:
    patch:
      operationId: updateAdminFlashSale
      tags: [admin]
      summary: Update flash sale
      parameters:
        - in: path
          name: saleId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashSaleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFlashSale"
    delete:
      operationId: deleteAdminFlashSale
      tags: [admin]
      summary: Delete flash sale
      parameters:
        - in: path
          name: saleId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFlashSale"
  /admin/flash-sales/{saleId}/products:
    post:
      operationId: addProductToFlashSale
      tags: [admin]
      summary: Add product to flash sale
      parameters:
        - in: path
          name: saleId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashSaleProductAdd"
      responses:
        "200":
          description: Updated sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFlashSale"

  # ─── ADMIN INVENTORY ──────────────────────────────────────────
...
components:
  schemas:
...
    Driver:
      type: object
      required: [id, name, phone, vehicleType, status, rating]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        avatar:
          type: ["string", "null"]
        vehicleType:
          type: string
          enum: [scooter, bicycle, car, van, truck]
        vehiclePlate:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, busy]
        rating:
          type: number
        totalDeliveries:
          type: integer
        wallet:
          type: number
    DriverUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        vehicleType:
          type: string
        vehiclePlate:
          type: string
...
    DriverStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [online, offline, busy]

    # ---- ANALYTICS ----
...
    FaqCategory:
      type: object
      required: [id, name, slug, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        displayOrder:
          type: integer
    FaqCategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        displayOrder:
          type: integer
    FaqItem:
      type: object
      required: [id, question, answer, isPublished, displayOrder]
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        displayOrder:
          type: integer
        isPublished:
          type: boolean
        createdAt:
          type: string
    FaqItemInput:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        displayOrder:
          type: integer
        isPublished:
          type: boolean

    # ─── BLOG ────────────────────────────────────────────────────────
...
    AdminFlashSale:
      type: object
      required: [id, title, endsAt, discountPercent, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        bannerImage:
          type: ["string", "null"]
        endsAt:
          type: string
        discountPercent:
          type: integer
        isActive:
          type: boolean
        productCount:
          type: integer
        createdAt:
          type: string
    FlashSaleInput:
      type: object
      required: [title, endsAt, discountPercent]
      properties:
        title:
          type: string
        description:
          type: string
        bannerImage:
          type: string
        endsAt:
          type: string
        discountPercent:
          type: integer
        isActive:
          type: boolean

    # ─── INVENTORY ───────────────────────────────────────────────────
...
    FlashSaleProductAdd:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer